backend_python/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
django-ratelimit>=3.0
mysqlclient>=2.1
pymongo>=4.3
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
pandas>=2.0
requests>=2.28
beautifulsoup4>=4.12
serpapi==0.1.5
python-dotenv>=1.0
pytest>=7.3
pytest-django>=4.8